requirements.txt
bson==0.5.10
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.18.0
requests==2.25.1
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
